Create RCB_and_playoffs
Sample Input 1 
3
4 10 8
3 6 1 
4 8 2 
Sample Output 1 
YES
NO
YES
Explanation
Test Case 1: There are 8 games remaining. Out of these 8 games, if RCB wins 2 games, loses 4 games and draws the remaining 2 games they will have a total of 10 points, this shows that it is possible for RCB to qualify for the playoffs.

Note: There can be many other combinations which will lead to RCB qualifying for the playoffs.

Test Case 2: There is no way such that RCB can qualify for the playoffs.

Test Case 3: If RCB wins all their remaining games, they will have end up with 8 points. (4+2∗2=8), hence it is possible for them to qualify for the playoffs.